backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
alembic==1.15.2
annotated-types==0.7.0
anyio==4.9.0
asyncer==0.0.8
attrs==25.3.0
backoff==2.2.1
bcrypt==4.3.0
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
cloudpickle==3.1.1
colorama==0.4.6
colorlog==6.9.0
cryptography==44.0.3
datasets==3.5.1
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
dspy==2.6.23
ecdsa==0.19.1
fastapi==0.115.12
filelock==3.18.0
frozenlist==1.6.0
fsspec==2025.3.0
greenlet==3.2.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.30.2
idna==3.10
importlib_metadata==8.7.0
Jinja2==3.1.6
jiter==0.9.0
joblib==1.5.0
json_repair==0.44.1
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
litellm==1.68.0
magicattr==0.1.6
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
multidict==6.4.3
multiprocess==0.70.16
numpy==2.2.5
openai==1.75.0
optuna==4.3.0
packaging==25.0
pandas==2.2.3
passlib==1.7.4
pdfminer.six==20250327
pdfplumber==0.11.6
pillow==11.2.1
propcache==0.3.1
pyarrow==20.0.0
pyasn1==0.4.8
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
Pygments==2.19.1
pypdfium2==4.30.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-jose==3.4.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rich==14.0.0
rpds-py==0.24.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.2
tenacity==9.1.2
tiktoken==0.9.0
tokenizers==0.21.1
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
ujson==5.10.0
urllib3==2.4.0
uvicorn==0.34.2
xxhash==3.5.0
yarl==1.20.0
zipp==3.21.0
